# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='700px' src='https://nitslaszlo.github.io/static/session/db_diagram_new.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://FarmShop.cyclic.app"
        }
    ],
    "paths": {
...
        "/partners": {
            "get": {
                "tags": [
                    "partners"
                ],
                "summary": "Összes partner lekérdezése.",
                "description": "Az összes partner adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/partner"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "partners"
                ],
                "summary": "Új partner mentése.",
                "description": "Új partner mentése a 'partner' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/partner"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett partnerel tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/partner"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "name": []
        }
    ],
    "components": {
        "schemas": {
...
            "partner": {
                "type": "object",
                "required": [
                    "title"
                ],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A partner azonosítója (ObjectID).",
                        "example": "d00000000000000000000001"
                    },
                    "content": {
                        "type": "string",
                        "description": "A partner tartalma.",
                        "example": "Sok féle termék közül lehet válogatni!"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "description": "A partner címe címe.",
                        "example": "Terméklesen"
                    }
                }
            },
...
            "ErrorSchema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error.message value"
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    }
                }
            }
        }
    }
}